Avoid throwing exceptions when using UDP socket
The `send` and `recv` methods threw when operating on an UDP socket.
The reason for that is not entirely clear, however we need to be able to send
and receive from an UDP socket. The operations work fine for UDP sockets
so the throws have been commented out.

Fixes https://github.com/PlasticSCM/UdtSharp/issues/4