Improve error handling (exception re-throwing)
This re-throw patterns in C# look weird, because when you create
a new exception and pass the previously-caught exception as an
innerException, normally you get all the information of all the
inner exceptions anyway when printing ex.ToString() (that is, not
only the stacktrace but all the exception messages of all the
inner exceptions).

But by looking at how UdtException was implemented, one can
see that the Message property of its base class was not being
set properly, hence the call to getErrorMessage() had to be
used in all possible places where it is thrown. With this
commit we fix this to map better with the C# conventions, and
instead of removing the re-throwing, we add more information
to the exception wrapper instead.